+++ php/src/Controller/DockerController.php
@@ -6,6 +6,7 @@ namespace AIO\Controller;
 use AIO\Container\Container;
 use AIO\Container\ContainerState;
 use AIO\ContainerDefinitionFetcher;
+use AIO\Desec\DesecManager;
 use AIO\Docker\DockerActionManager;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
@@ -19,7 +20,7 @@ readonly class DockerController {
         private DockerActionManager           $dockerActionManager,
         private ContainerDefinitionFetcher    $containerDefinitionFetcher,
         private ConfigurationManager $configurationManager,
-        private DesecController $desecController,
+        private DesecManager $desecManager,
     ) {
     }
 
@@ -265,7 +266,7 @@ readonly class DockerController {
         $this->StopDomaincheckContainer();
 
         // Refresh the deSEC DNS record with the current public IP before starting containers
-        $this->desecController->updateIpIfDesecDomain();
+        $this->desecManager->updateIpIfDesecDomain();
 
         $id = self::TOP_CONTAINER;
